fix: feedback + cleaner code
