+++ apps/ledger-live-mobile/scripts/sync-families-dispatch.mjs
@@ -28,7 +28,7 @@ const targets = [
   "AccountBalanceSummaryFooter",
   "SubAccountList",
   "Confirmation",
-  "PostConfirmationAlert",
+  "ReceiveConfirmationPostAlert",
   "ConnectDevice",
   "NoAssociatedAccounts",
 ];
